Split release steps by domain
